double to float

+++ src/engine/geom/shape/Circle.java
@@ -4,36 +4,36 @@ import engine.physics2d.Vector;
 import java.awt.geom.Ellipse2D;
 
 public class Circle extends Shape {
-	private double radius;
+	private float radius;
 	private Vector origin;
 
-	public double getRadius() {
+	public float getRadius() {
 		return radius;
 	}
 
 	@Override
-	public void rotate(double theta) {
+	public void rotate(float theta) {
 		rotation -= theta;
 	}
 
 	@Override
-	public void translateX(double dx) {
+	public void translateX(float dx) {
 		origin.translateX(dx);
 	}
 
 	@Override
-	public void translateY(double dy) {
+	public void translateY(float dy) {
 		origin.translateY(dy);
 	}
 
 	@Override
 	public java.awt.Shape stroke() {
-		return new Ellipse2D.Double(origin.getX(), origin.getY(), radius, radius);
+		return new Ellipse2D.Float(origin.getX(), origin.getY(), radius, radius);
 	}
 
 	@Override
-	public double area() {
-		return Math.PI * radius * radius;
+	public float area() {
+		return (float) (Math.PI * radius * radius);
 	}
 
 	@Override
@@ -47,9 +47,9 @@ public class Circle extends Shape {
 	}
 
 	@Override
-	public boolean contains(double x, double y) {
-		double dx = origin.getX() - x * origin.getX() - x;
-		double dy = origin.getY() - y * origin.getY() - y;
+	public boolean contains(float x, float y) {
+		float dx = origin.getX() - x * origin.getX() - x;
+		float dy = origin.getY() - y * origin.getY() - y;
    		return dx + dy <= radius * radius;
 	}
 }
