+++ src/engine/physics2d/Collisions.java
@@ -39,18 +39,18 @@ public class Collisions {
 			Collider A = contact.colliderA;
 			Collider B = contact.colliderB;
 
-			double va = A.velocity.dot(normal), vb = B.velocity.dot(normal);
-			double bouciness = Math.max(A.restitution, B.restitution);
-			double friction = Math.min(A.friction, B.friction) * Vector.sub(A.velocity, B.velocity).dot(tangent);
+			float va = A.velocity.dot(normal), vb = B.velocity.dot(normal);
+			float bouciness = Math.max(A.restitution, B.restitution);
+			float friction = Math.min(A.friction, B.friction) * Vector.sub(A.velocity, B.velocity).dot(tangent);
 
 			if (A.isDynamic()) {
-				double vf = bouciness * (2 * A.im * vb + (B.im - A.im) * va) / (A.im + B.im);
-				A.updateImpulse(Vector.scale(normal, (Math.abs(vf) < 0.01 ? 0 : vf) - va));
+				float vf = bouciness * (2 * A.im * vb + (B.im - A.im) * va) / (A.im + B.im);
+				A.updateImpulse(Vector.scale(normal, (Math.abs(vf) < 0.01f ? 0 : vf) - va));
 				A.updateImpulse(Vector.scale(tangent, friction * -A.im));
 			}
 			if (B.isDynamic()) {
-				double vf = bouciness * (2 * B.im * va + (A.im - B.im) * vb) / (A.im + B.im);
-				B.updateImpulse(Vector.scale(normal, (Math.abs(vf) < 0.01 ? 0 : vf) - vb));
+				float vf = bouciness * (2 * B.im * va + (A.im - B.im) * vb) / (A.im + B.im);
+				B.updateImpulse(Vector.scale(normal, (Math.abs(vf) < 0.01f ? 0 : vf) - vb));
 				B.updateImpulse(Vector.scale(tangent, friction * B.im));
 			}
 		}
@@ -59,20 +59,20 @@ public class Collisions {
 	/**
 	 * The penetration allowance.
 	 */
-	private static final double PENETRATION_ALLOWANCE = 0.05;
+	private static final float PENETRATION_ALLOWANCE = 0.05f;
 
 	/**
 	 * The penetration percentage to correct
 	 */
-	private static final double PENETRATION_CORRETION = 0.9;
+	private static final float PENETRATION_CORRETION = 0.9f;
 
 	/**
 	 * This applies a positional correction for each contact.
 	 */
 	public static void correction() {
 		for (Manifold contact : contacts) {
-			if (contact.colliderA.im + contact.colliderB.im > 0.0) {
-				double correction =  Math.max(contact.penetration - PENETRATION_ALLOWANCE, 0.0) / (contact.colliderA.im + contact.colliderB.im) * PENETRATION_CORRETION;
+			if (contact.colliderA.im + contact.colliderB.im > 0) {
+				float correction =  Math.max(contact.penetration - PENETRATION_ALLOWANCE, 0) / (contact.colliderA.im + contact.colliderB.im) * PENETRATION_CORRETION;
 				if (contact.colliderA.isDynamic())
 					contact.colliderA.getBounds().translate(Vector.scale(contact.normal, contact.colliderA.im * correction));
 				if (contact.colliderB.isDynamic())
