+++ src/game/level/TileTest.java
@@ -5,13 +5,13 @@ import java.awt.Color;
 import engine.scene.entity.Tile;
 
 public class TileTest extends Tile {
-	public TileTest(String name, double x, double y, double width, double height, int layer) {
+	public TileTest(String name, float x, float y, float width, float height, int layer) {
 		super(name, x, y, width, height, layer);
 		solid = true;
 	}
 
 	@Override
-	public void update(double dt) {}
+	public void update(float dt) {}
 
 	@Override
 	protected void draw(Graphics2D graphics) {
