+++ src/sandbox/collission/PhyEntity.java
@@ -14,19 +14,19 @@ public class PhyEntity extends Entity {
 	 * @param width  the width of the bounding {@code Rectangle}
 	 * @param height the height of the bounding {@code Rectangle}
 	 */
-	public PhyEntity(double x, double y, double width, double height, int zindex) {
+	public PhyEntity(float x, float y, float width, float height, int zindex) {
 		super(x, y, width, height, zindex);
-		im = 0.02;
+		im = 0.02f;
 	}
 
-	public PhyEntity(double x, double y, double width, double height, int zindex, boolean b) {
+	public PhyEntity(float x, float y, float width, float height, int zindex, boolean b) {
 		super(x, y, width, height, zindex);
 		type = CollisionType.KINEMATIC;
-		im = 0.0;
+		im = 0;
 	}
 
 	@Override
-	public void update(double dt) {
+	public void update(float dt) {
 		bounds.translate(Vector.scale(velocity, dt));
 	}
